pc_control_system/windows_client/requirements.txt
requests==2.31.0
websocket-client==1.6.4
psutil==5.9.6
python-dotenv==1.0.0 